cointracker/requirements.txt
fastapi>=0.78.0
httpx>=0.23.0
instructor>=0.4.2
openai>=1.3.6
parso>=0.5.1
python-dotenv>=0.19.0
pydantic>=2.5
requests>=2.3.1
uvicorn[standard]>=0.18.2
sqlalchemy==1.4.3  # SQLAlchemy for ORM
databases==0.6.0  # Async database support
alembic==1.7.4  # Database migrations
sqlalchemy-repr==0.1.0  # Better repr for SQLAlchemy models
psycopg2==2.9.9  # PostgreSQL database driver
black